Added field RemoteFileID to RemoteFileInfo.cs. Updated EnumerateFiles in B2RemoteStorage.cs to propagate the FileId field returned from the enumeration from B2 into the RemoteFileID field of the RemoteFileInfo objects it yields. Updated OutputFileListBase.cs in DeltaQ.RTB.Restore to supply a dummy fileID value when constructing RemoteFileInfo. Updated ZFS.cs and its subclass ZFSSnapshot.cs to take an IErrorLogger on the constructor. Updated all call sites to provide one. Updated automated tests to include an assertion that the error logger did not receive any LogError calls during each test. Updated ExecuteZFSCommand in ZFS.cs to return the exit code of the invoked zfs child process. Updated the Dispose method in ZFSSnapshot.cs to catch errors and implement a delay retry loop, only reporting the failure to destroy a ZFS snapshot if it fails 5 times in 50 seconds. Renamed B2RemoteStorage.GetFileIdByName to GetFileIDByName for consistency. Updated RetryIfNoTomesAreAvailable in B2RemoteStorage.cs to be resilient to the action result not having an HttpResponse at all. Removed the unused bucketId parameter from DownloadFileBytes in B2RemoteStorage.cs. Added methods DownloadFileByID and DeleteFileByID to IRemoteStorage.cs and implemented them in B2RemoteStorage.cs. Fixed EnumerateFiles so that it doesn't strip the leading '/' character if the supplied pathPrefix is empty. Updated EnumerateFiles to use ListFileVersionRequest instead of ListFileNamesRequest so that it can identify duplicates. Updated ShowWindow in App.axaml.cs to use MainWindow.Activate if the window is already visible. Added switches /LISTGHOSTSTATEFILES, /LISTFILESWITHMULTIPLEVERSIONS and /REMOVEFILESNOTINCACHE to CommandLineArguments.cs for DeltaQ.RTB.Tweaker. Implemented them in Program.cs.

+++ DeltaQ.RTB.Tweaker/Program.cs
@@ -1,5 +1,7 @@
 using System;
+using System.Collections.Generic;
 using System.IO;
+using System.Text;
 using System.Xml.Serialization;
 
 using CancellationToken = System.Threading.CancellationToken;
@@ -8,7 +10,8 @@ using Autofac;
 
 using Bytewizer.Backblaze.Client;
 
-using DeltaQ.RTB;
+using DeltaQ.RTB.Bridge.Notifications;
+using DeltaQ.RTB.Diagnostics;
 using DeltaQ.RTB.StateCache;
 using DeltaQ.RTB.Storage;
 using DeltaQ.RTB.Utility;
@@ -34,6 +37,8 @@ namespace DeltaQ.RTB.Tweaker
 
 			builder.AddBackblazeAgent(backblazeAgentOptions);
 
+			builder.RegisterType<ErrorLogger>().AsImplementedInterfaces().SingleInstance();
+			builder.RegisterType<NotificationBus>().AsImplementedInterfaces().SingleInstance();
 			builder.RegisterType<B2RemoteStorage>().AsImplementedInterfaces().SingleInstance();
 			builder.RegisterType<RemoteFileStateCache>().AsImplementedInterfaces().SingleInstance();
 			builder.RegisterType<RemoteFileStateCacheRemoteStorage>().AsImplementedInterfaces().SingleInstance();
@@ -79,7 +84,7 @@ namespace DeltaQ.RTB.Tweaker
 					System.Threading.CancellationToken.None);
 			}
 
-			if (arguments.DeleteGhostStateFiles)
+			if (arguments.ListGhostStateFiles || arguments.DeleteGhostStateFiles)
 			{
 				Console.WriteLine("Enumerating remote file state cache files on the server...");
 
@@ -95,8 +100,13 @@ namespace DeltaQ.RTB.Tweaker
 						Console.WriteLine("Exists locally, leaving untouched");
 					else
 					{
-						Console.WriteLine("Does not exist locally, sending a deletion to the server");
-						remoteStorage.DeleteFileDirect(serverPath, System.Threading.CancellationToken.None);
+						if (arguments.DeleteGhostStateFiles)
+						{
+							Console.WriteLine("Does not exist locally, sending a deletion to the server");
+							remoteStorage.DeleteFileDirect(serverPath, System.Threading.CancellationToken.None);
+						}
+						else
+							Console.WriteLine("File does not exist locally and should have been deleted from remote storage.");
 					}
 				}
 			}
@@ -120,6 +130,118 @@ namespace DeltaQ.RTB.Tweaker
 						Console.WriteLine("    Cancelling large file");
 
 						remoteStorage.CancelUnfinishedFile(unfinishedFile.Path, CancellationToken.None);
+					}
+				}
+			}
+
+			if (arguments.ListFilesWithMultipleVersions)
+			{
+				Console.WriteLine("Enumerating files...");
+
+				var paths = new HashSet<string>();
+				var duplicatePaths = new HashSet<string>();
+
+				foreach (var file in remoteStorage.EnumerateFiles("/content/", recursive: true))
+				{
+					string filePath = file.Path;
+
+					if (!filePath.StartsWith("/"))
+						filePath = '/' + filePath;
+
+					if (!paths.Add(filePath))
+					{
+						if (duplicatePaths.Add(filePath))
+						{
+							if (duplicatePaths.Count == 1)
+								Console.WriteLine("Files with multiple versions in remote storage:");
+
+							Console.WriteLine("- {0}", filePath);
+						}
+					}
+
+					if ((paths.Count & 1023) == 0)
+					{
+						Console.Write(paths.Count);
+
+						if (Console.IsOutputRedirected)
+							Console.WriteLine();
+						else
+							Console.CursorLeft = 0;
+					}
+				}
+			}
+
+			if (arguments.RemoveIncorrectFiles)
+			{
+				Console.WriteLine("Enumerating files...");
+
+				var fileInfoByContentKey = new Dictionary<string, RemoteFileInfo>();
+				var pathByContentKey = new Dictionary<string, string>();
+
+				foreach (var file in remoteStorage.EnumerateFiles("", recursive: true))
+				{
+					string filePath = file.Path;
+
+					if (filePath.StartsWith("/state/"))
+						continue;
+
+					if (filePath.StartsWith("/content/"))
+					{
+						var buffer = new MemoryStream();
+
+						remoteStorage.DownloadFileDirect(filePath, buffer, CancellationToken.None);
+
+						string contentKey = Encoding.UTF8.GetString(buffer.ToArray());
+
+						pathByContentKey[contentKey] = filePath;
+					}
+					else
+						fileInfoByContentKey[filePath] = file;
+
+					if ((pathByContentKey.Count & 31) == 0)
+					{
+						Console.Write(pathByContentKey.Count + fileInfoByContentKey.Count);
+
+						if (Console.IsOutputRedirected)
+							Console.WriteLine();
+						else
+							Console.CursorLeft = 0;
+					}
+				}
+
+				Console.WriteLine("Enumerated {0} files", pathByContentKey.Count + fileInfoByContentKey.Count);
+				Console.WriteLine("Loading the Remote File State Cache...");
+
+				var remoteFileStateCache = container.Resolve<IRemoteFileStateCache>();
+
+				remoteFileStateCache.LoadCache();
+
+				foreach (var pathMapping in pathByContentKey)
+				{
+					string contentKey = pathMapping.Key;
+					string path = pathMapping.Value;
+
+					if (!fileInfoByContentKey.TryGetValue(contentKey, out var fileInfo))
+					{
+						Console.WriteLine("File pointer: {0}", path);
+						Console.WriteLine("... points at content key: {0}", contentKey);
+						Console.WriteLine("... but no content by that content key exists.");
+					}
+					else
+					{
+						var fileState = remoteFileStateCache.GetFileState(path);
+
+						if (fileState == null)
+						{
+							Console.WriteLine("File pointer: {0}", path);
+							Console.WriteLine("... is not in the Remote File State Cache");
+						}
+						else if (fileInfo.FileSize != fileState.FileSize)
+						{
+							Console.WriteLine("File pointer: {0}", path);
+							Console.WriteLine("... points at content with size {0:#,##0}", fileInfo.FileSize);
+							Console.WriteLine("... but the Remote File State Cache expects size {0:#,##0}", fileState.FileSize);
+						}
 					}
 				}
 			}
